+++ CHANGELOG.md
@@ -3,6 +3,8 @@
   - fixed serialisation for  svm
   - added tfidf->dense
   - added :tf-map-handler-fn to support pruning of the terms for tfidf
+  - metamorph'ed  bow->tfidf reuses tf-map from :fit in :transform
+
 
 ## v6.2.585
 - added LDA QDA RDA FLD]